servers/bionic/requirements.txt
flask>=2.0.0
python-docx>=0.8.11
werkzeug>=2.0.0
google-genai>=1.0.0
